[tools] Move the tools lib into its own directory
