+++ src/bench/rtree.rs
@@ -1,5 +1,5 @@
 use crate::fixture::*;
-use crate::pdata::rtree::*;
+use crate::tools::pdata::rtree::*;
 use crate::stubs::*;
 use crate::test_runner::*;
 use crate::tests::rtree::*;
