+++ src/lib.rs
@@ -16,7 +16,7 @@ pub mod emulator;
 pub mod fixture;
 pub mod guest;
 pub mod path_formatter;
-pub mod pdata;
+pub mod tools;
 pub mod primitive;
 pub mod stats;
 pub mod stubs;
